requirements.txt
requests
reportlab
python-dotenv
openai
google-generativeai
anthropic
python-owasp-zap-v2.4
